# client_java/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: iratefinan@etu.uqac.ca
    name: SBIR Team
  description: |
    API REST pour le système de gestion de tickets développé en Java pour le cours 6GEI311, Architecture Logicielle

    Auteurs: Il'aina Ratefinanahary et Samuel Brassard, UQAC Automne 2025
  license:
    name: Usage Éducatif Uniquement
    url: https://github.com/TechSamQC/Lab2-GEI311-SBIR
  title: "Système de Gestion de Tickets (SBIR), API REST"
  version: 4.0.0
servers:
- description: Serveur de développement local
  url: http://localhost:8080/api/v1
...
paths:
  /users:
    get:
      description: Retourne la liste complète des utilisateurs du système
      operationId: getAllUsers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/UserDTO'
                type: array
          description: Liste des utilisateurs récupérée avec succès
        "500":
          content:
            application/json:
              example:
                error: INTERNAL_ERROR
                message: Une erreur inattendue s'est produite
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Erreur interne du serveur
      summary: Liste tous les utilisateurs
      tags:
      - users
      x-accepts:
      - application/json
    post:
      description: Crée un nouvel utilisateur dans le système
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        description: Données de l'utilisateur à créer
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: Utilisateur créé avec succès
        "400":
          content:
            application/json:
              example:
                error: VALIDATION_ERROR
                message: Le titre du ticket ne peut pas être vide
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Requête invalide (validation échouée)
        "500":
          content:
            application/json:
              example:
                error: INTERNAL_ERROR
                message: Une erreur inattendue s'est produite
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Erreur interne du serveur
      summary: Crée un nouvel utilisateur
      tags:
      - users
      x-content-type: application/json
      x-accepts:
      - application/json
  /users/{userId}:
    delete:
      description: Supprime un utilisateur du système
      operationId: deleteUser
      parameters:
      - description: Identifiant unique de l'utilisateur
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int32
          type: integer
        style: simple
      responses:
        "204":
          description: Utilisateur supprimé avec succès
        "404":
          content:
            application/json:
              example:
                error: NOT_FOUND
                message: "Ticket #9999 introuvable"
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource non trouvée
        "403":
          content:
            application/json:
              example:
                error: FORBIDDEN
                message: Permissions insuffisantes pour cette opération
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Accès refusé (permissions insuffisantes)
        "500":
          content:
            application/json:
              example:
                error: INTERNAL_ERROR
                message: Une erreur inattendue s'est produite
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Erreur interne du serveur
      summary: Supprime un utilisateur
      tags:
      - users
      x-accepts:
      - application/json
    get:
      description: Retourne les détails d'un utilisateur par son ID
      operationId: getUserById
      parameters:
      - description: Identifiant unique de l'utilisateur
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int32
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: Utilisateur trouvé
        "404":
          content:
            application/json:
              example:
                error: NOT_FOUND
                message: "Ticket #9999 introuvable"
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource non trouvée
        "500":
          content:
            application/json:
              example:
                error: INTERNAL_ERROR
                message: Une erreur inattendue s'est produite
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Erreur interne du serveur
      summary: Récupère un utilisateur spécifique
      tags:
      - users
      x-accepts:
      - application/json
    put:
      description: Met à jour les informations d'un utilisateur existant
      operationId: updateUser
      parameters:
      - description: Identifiant unique de l'utilisateur
        explode: false
        in: path
        name: userId
        required: true
        schema:
          example: 1
          format: int32
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDTO'
          description: Utilisateur mis à jour avec succès
        "400":
          content:
            application/json:
              example:
                error: VALIDATION_ERROR
                message: Le titre du ticket ne peut pas être vide
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Requête invalide (validation échouée)
        "404":
          content:
            application/json:
              example:
                error: NOT_FOUND
                message: "Ticket #9999 introuvable"
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Ressource non trouvée
        "500":
          content:
            application/json:
              example:
                error: INTERNAL_ERROR
                message: Une erreur inattendue s'est produite
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Erreur interne du serveur
      summary: Met à jour un utilisateur
      tags:
      - users
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
...
  schemas:
    UserDTO:
      description: Représentation d'un utilisateur du système
      example:
        role: DEVELOPER
        name: Utilisateur1
        isAdmin: false
        userID: 1
        email: utilisateur1@uqac.ca
      properties:
        userID:
          description: Identifiant unique de l'utilisateur
          example: 1
          format: int32
          type: integer
        name:
          description: Nom complet de l'utilisateur
          example: Utilisateur1
          maxLength: 100
          minLength: 1
          type: string
        email:
          description: Adresse email de l'utilisateur
          example: utilisateur1@uqac.ca
          format: email
          type: string
        role:
          description: Rôle de l'utilisateur dans le système
          enum:
          - ADMIN
          - DEVELOPER
          - USER
          example: DEVELOPER
          type: string
        isAdmin:
          description: Indique si l'utilisateur a les privilèges admin
          example: false
          type: boolean
      required:
      - email
      - isAdmin
      - name
      - role
      - userID
      type: object
...
    UserCreate:
      description: Requête de création d'un nouvel utilisateur
      example:
        role: DEVELOPER
        name: William Derreck
        email: william.derreck@uqac.ca
      properties:
        name:
          description: Nom complet
          example: William Derreck
          type: string
        email:
          description: Adresse email
          example: william.derreck@uqac.ca
          format: email
          type: string
        role:
          description: Rôle
          enum:
          - ADMIN
          - DEVELOPER
          - USER
          example: DEVELOPER
          type: string
      required:
      - email
      - name
      - role
      type: object
    UserUpdate:
      description: Requête de modification d'un utilisateur
      example:
        role: ADMIN
        name: William Derreck
        email: william.derreck@uqac.ca
      properties:
        name:
          example: William Derreck
          type: string
        email:
          example: william.derreck@uqac.ca
          format: email
          type: string
        role:
          enum:
          - ADMIN
          - DEVELOPER
          - USER
          type: string
      type: object
    ErrorResponse:
      description: Réponse d'erreur standardisée
      example:
        code: 400
        details: details
        error: INVALID_TRANSITION
        message: "Transition invalide : Ouvert -> Termine. Transitions autorisées\
          \ : ASSIGNE, FERME"
      properties:
        error:
          description: Code d'erreur
          example: INVALID_TRANSITION
          type: string
        code:
          description: Code d'erreur numérique
          example: 400
          format: int32
          type: integer
        message:
          description: Message d'erreur détaillé
          example: "Transition invalide : Ouvert -> Termine. Transitions autorisé\
            es : ASSIGNE, FERME"
          type: string
        details:
          description: Détails supplémentaires sur l'erreur
          type: string
      required:
      - error
      - message
      type: object
...