# server_springboot/src/main/java/org/openapitools/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ErrorResponse;
import org.openapitools.model.UserCreate;
import org.openapitools.model.UserDTO;
import org.openapitools.model.UserUpdate;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import javax.validation.Valid;
import java.util.List;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-18T07:20:00.634681-05:00[America/Toronto]", comments = "Generator version: 7.10.0")
@Validated
@Tag(name = "users", description = "Gestion des utilisateurs")
public interface UsersApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /users : Crée un nouvel utilisateur
     * Crée un nouvel utilisateur dans le système
     *
     * @param userCreate Données de l&#39;utilisateur à créer (required)
     * @return Utilisateur créé avec succès (status code 201)
     *         or Requête invalide (validation échouée) (status code 400)
     *         or Erreur interne du serveur (status code 500)
     */
    @Operation(
        operationId = "createUser",
        summary = "Crée un nouvel utilisateur",
        description = "Crée un nouvel utilisateur dans le système",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Utilisateur créé avec succès", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide (validation échouée)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erreur interne du serveur", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<UserDTO> createUser(
        @Parameter(name = "UserCreate", description = "Données de l'utilisateur à créer", required = true) @Valid @RequestBody UserCreate userCreate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"role\" : \"DEVELOPER\", \"name\" : \"Utilisateur1\", \"isAdmin\" : false, \"userID\" : 1, \"email\" : \"utilisateur1@uqac.ca\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Validation. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Validation. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /users/{userId} : Supprime un utilisateur
     * Supprime un utilisateur du système
     *
     * @param userId Identifiant unique de l&#39;utilisateur (required)
     * @return Utilisateur supprimé avec succès (status code 204)
     *         or Ressource non trouvée (status code 404)
     *         or Accès refusé (permissions insuffisantes) (status code 403)
     *         or Erreur interne du serveur (status code 500)
     */
    @Operation(
        operationId = "deleteUser",
        summary = "Supprime un utilisateur",
        description = "Supprime un utilisateur du système",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Utilisateur supprimé avec succès"),
            @ApiResponse(responseCode = "404", description = "Ressource non trouvée", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "403", description = "Accès refusé (permissions insuffisantes)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erreur interne du serveur", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/users/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deleteUser(
        @Parameter(name = "userId", description = "Identifiant unique de l'utilisateur", required = true, in = ParameterIn.PATH) @PathVariable("userId") Integer userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Validation. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Validation. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users : Liste tous les utilisateurs
     * Retourne la liste complète des utilisateurs du système
     *
     * @return Liste des utilisateurs récupérée avec succès (status code 200)
     *         or Erreur interne du serveur (status code 500)
     */
    @Operation(
        operationId = "getAllUsers",
        summary = "Liste tous les utilisateurs",
        description = "Retourne la liste complète des utilisateurs du système",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Liste des utilisateurs récupérée avec succès", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = UserDTO.class)))
            }),
            @ApiResponse(responseCode = "500", description = "Erreur interne du serveur", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<UserDTO>> getAllUsers(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"role\" : \"DEVELOPER\", \"name\" : \"Utilisateur1\", \"isAdmin\" : false, \"userID\" : 1, \"email\" : \"utilisateur1@uqac.ca\" }, { \"role\" : \"DEVELOPER\", \"name\" : \"Utilisateur1\", \"isAdmin\" : false, \"userID\" : 1, \"email\" : \"utilisateur1@uqac.ca\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /users/{userId} : Récupère un utilisateur spécifique
     * Retourne les détails d&#39;un utilisateur par son ID
     *
     * @param userId Identifiant unique de l&#39;utilisateur (required)
     * @return Utilisateur trouvé (status code 200)
     *         or Ressource non trouvée (status code 404)
     *         or Erreur interne du serveur (status code 500)
     */
    @Operation(
        operationId = "getUserById",
        summary = "Récupère un utilisateur spécifique",
        description = "Retourne les détails d'un utilisateur par son ID",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Utilisateur trouvé", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTO.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource non trouvée", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erreur interne du serveur", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{userId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<UserDTO> getUserById(
        @Parameter(name = "userId", description = "Identifiant unique de l'utilisateur", required = true, in = ParameterIn.PATH) @PathVariable("userId") Integer userId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"role\" : \"DEVELOPER\", \"name\" : \"Utilisateur1\", \"isAdmin\" : false, \"userID\" : 1, \"email\" : \"utilisateur1@uqac.ca\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /users/{userId} : Met à jour un utilisateur
     * Met à jour les informations d&#39;un utilisateur existant
     *
     * @param userId Identifiant unique de l&#39;utilisateur (required)
     * @param userUpdate  (required)
     * @return Utilisateur mis à jour avec succès (status code 200)
     *         or Requête invalide (validation échouée) (status code 400)
     *         or Ressource non trouvée (status code 404)
     *         or Erreur interne du serveur (status code 500)
     */
    @Operation(
        operationId = "updateUser",
        summary = "Met à jour un utilisateur",
        description = "Met à jour les informations d'un utilisateur existant",
        tags = { "users" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Utilisateur mis à jour avec succès", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDTO.class))
            }),
            @ApiResponse(responseCode = "400", description = "Requête invalide (validation échouée)", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Ressource non trouvée", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Erreur interne du serveur", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/users/{userId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<UserDTO> updateUser(
        @Parameter(name = "userId", description = "Identifiant unique de l'utilisateur", required = true, in = ParameterIn.PATH) @PathVariable("userId") Integer userId,
        @Parameter(name = "UserUpdate", description = "", required = true) @Valid @RequestBody UserUpdate userUpdate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"role\" : \"DEVELOPER\", \"name\" : \"Utilisateur1\", \"isAdmin\" : false, \"userID\" : 1, \"email\" : \"utilisateur1@uqac.ca\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"code\" : 400, \"details\" : \"details\", \"error\" : \"INVALID_TRANSITION\", \"message\" : \"Transition invalide : Ouvert -> Termine. Transitions autorisées : ASSIGNE, FERME\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}